Overwrite path_helper file on install.
If the package's installer is run multiple times, the installation
path would be appended for every time the package had been run. This
change overwrites the file initially, and if subsequent paths need
to be added, then those can be appended. This ensures the "dotnet"
path helper file is clean for each install.

Fixes #18.