Adding support for symlinking libssl & libcrypto
Recently, Homebrew has changed its policies to dissalow the --force
linking of OpenSSL. Because we need OpenSSL in the loader path, we had
several options and the best one we determined to use, given all of the
scenarios we know, was to add symlinks to versions that we require to
/usr/local/lib.

This change makes that operation automatic when installing from the
PKG as part of its postinstall script. However, the user would still
need to install OpenSSL through brew in order for this to work.