Mark Microsoft.NETCore.App 1.0.1 stable
Also update the package build to only build Microsoft.NETCore.App
as it is the only package we are shipping.

Build all packages in core-setup.

Core-setup build depends on all packages so we need to build them
all, however that will likely cause package publishing to fail
for the ones we have already published stable. We will need to
watch out for that with our official build.

asd