Add checked-in build definitions
The Windows signing build definitions will be added later, because the
VSTS instance where we want to run the builds doesn't support the
current signing process, so we're working on modifying our signing
process.

In a separate PR, I plan to refactor the Linux build definitions into a
smaller number of build definitions.