Fix yum package conflict in Docker image
When running the Core-Setup builds on the new build agents, the CentOS
build failed with the following error:

2017-01-11T18:41:24.6447580Z Step 4 : RUN yum -q -y install unzip libunwind gettext libcurl-devel openssl-devel zlib libicu-devel
2017-01-11T18:41:24.7898330Z  ---> Running in cb2648bea479
2017-01-11T18:41:32.7647660Z  You could try using --skip-broken to work around the problem
2017-01-11T18:41:32.7663610Z _[91mError: libselinux conflicts with fakesystemd-1-17.el7.centos.noarch

Swapping the the "fakesystemd" package with the real "systemd" package
fixed the issue, which was recommended in http://stackoverflow.com/a/36632668.

This issue hasn't been seen on the official builds because the build
agents in that pool have the Docker image cached, and they are using an
older version of the centos:7.1.1503 image which appears not to have
this issue.