Add AppleCrypto shim as a direct NETCore.App dependency for servicing
runtime.native.System.Security.Cryptography.Apple is an indirect
dependency, so building new versions of it out of CoreFX doesn't result
in the tooling pushing it to a higher minimum version when we service it.

This change adds the oldest 4.3.1-servicing package as a direct dependency,
and the version will roll forward via the normal upgrade-to-newer-builds
process.