Remove runtimemodel reference
I built core-setup locally, and didn't see runtimemodel package any more
with current version 4.3.0-preview2-4095. It is possible that it might
not be needed any longer.

Remove it first and verify it all works. If there is still runtime
errors then we may need to explicitly add a package reference to it