[1.1] Fix Package restore error for new distros
The new distros added for 1.1 currently have an issue where the tools restoration for the build/packaging stage is pulling in an old version of System.Security.Cryptograph.X509Certifications (and a few others) that directly reference the runtime-specific meta-package component runtime.native.System.Security.Cryptography.OpenSsl. This is an issue because the 4.3.0 version being restored does not include support for the new distros, so we're seeing dll missing errors on those platforms.

There are a few possible long-term solutions to this problem from the CoreFX repo:
1. Ship a new version of NETStandard.Library that contains an updated version of the managed packages that reference the Crypto runtime component.
2. Modify Microsoft.NETCore.Targets to include runtime package references for the crypto native packages. This would resolve the issue for any project on a new distro that references both the Targets package as well as the NetStandardLibrary package, as the newer runtime specific package would win out during package restoration.

This commit doesn't aim to do those things but works around the issue in core-setup by directly referencing the OpenSsl native package from the projects that actually use the cryptographic methods that rely upon the OpenSsl native package.