Avoid logging to stderr if core_servicing location is empty (#4110) (#4117)
Addresses #3852 by not logging to stderr when the call to realpath() fails when resolving the servicing location.

The issue is the servicing location "\Program Files (x86)" does not exist, and the call to realpath() fails. 
Note if that directory is manually created, the error also goes away.

Verified by reproducing the original problem, and then replacing the hostpolicy.dll in the Docker container with 
the "fixed" version. Then the "Error resolving full path []" message was not displayed.