Pass STARTUP_HOOKS property key to coreclr_initialize
STARTUP_HOOKS is now passed to the clr during initialization, instead
of using create_delegate in the host. This will allow the clr to call
the startup hooks closer to the execution of the main method, allowing
the threading apartment state to be set based on the main method's
attributes.