[release/2.2] Startup hook host functionality (#4574)
* Add startup hook design document (#4421)

* Add startup hook design document

This adds a proposal for an early startup hook that allows managed
code to run before the entry point.

* Change env var behavior and clarify version constraints

* Fix formatting, specify Initialize() signature

* Update spec for multiple hooks in environment variable

- Move the configuration to a single environment variable,
  DOTNET_STARTUP_HOOKS
- Introduce syntax to allow multiple hooks to be defined
- Remove .runtimeconfig file options
  This will only support the environment variable, since
  this is a low-level feature that should only be used when implicit
  injection of code is desired.
- Update implementation details

* Remove default type name and other fixes

- Require an explicit type name for each startup hook
- Fix Initialize() signature (it must be static)
- Clarify that the Initialize() methods are called synchronously

* Require full path, clarify syntax, and change TPA behavior

- Relative paths to the assemblies are not allowed
- Use platform-specific path separator
- Clarify restrictions on the syntax (no empty entries, trailing separator)
- Startup assemblies will not be placed on TPA list
  Instead, they will be loaded directly using LoadFromAssemblyPath in
  the implementation.

* Add guidance, error behavior, and thread behavior

* Remove specific error text, fix typos and formatting

* Remove type name from synax

The type must now be called StartupHook, and can not be set by the
environment variable. The descriptions and examples were updated to
reflect this.

Also:
- add a few notes about the visibility of this type
- describe behavior of uncaught exceptions in startup hooks
- give example with platform-specific separators
- fix some typos

* Change startup hook to run closer to Main

Instead of the host creating a delegate and inoking it,
ExecuteMainMethod in the runtime will call the System.Private.CoreLib
function that processes startup hooks. This allows the thread
apartment state to be set based on any attributes in the Main method,
before startup hooks execute.

* Clarify apartment threading state behavior

* Specify the property name passed to the runtime

* Add host startup hook and tests (#4465)

* Expose coreclr_create_delegate function

* Add startup hook call

Read DOTNET_STARTUP_HOOKS environment variable, and pass it along to a
private function in
System.Private.CoreLib (StartupHookProvider.ProcessStartupHookS).

* Add startup hook tests

* Add some more syntax checks

* Reorder the tests

* Add test for exclamation mark in assembly path

* Fix error behavior for missing methods

MissingMethodException is now only thrown when the Initialize method
doesn't exist at all. This also reorganizes the existing tests for the
invalid signature, and adds some cases.

* Change test behavior to not expect eager file exists check

* Expect full paths for startup hook.

Fixes this in existing tests, and adds a new check for the error
behavior when a relative path is passed.

* Add test for startup hook trace output

* Update tests to use new syntax (without type name)

* Pass STARTUP_HOOKS property key to coreclr_initialize

STARTUP_HOOKS is now passed to the clr during initialization, instead
of using create_delegate in the host. This will allow the clr to call
the startup hooks closer to the execution of the main method, allowing
the threading apartment state to be set based on the main method's
attributes.

* Allow non-public Initialize method

* Remove coreclr::create_delegate

* Add comments clarifying purpose of startup hook test projects

* Add test with ALC assembly resolve event

The startup hook will inject a dependency into the app.

* Fix the ALC startup hook testcase

- Don't set ReferenceOutputAssembly in the projectreference, because
  that results in it not being published.
- Dispose the new projects
- Don't throw exceptions from the resolver, since they are swallowed.