Allow missing assembly dependencies for components.
When resolving assembly dependencies for components, don't fail if the assembly doesn't exist on the disk (but it's still present in the .deps.json as a runtime asset).
For dynamically loaded components this behavior makes sense as such components are typically distributed as just the necessary pieces, with everything else provided by the app.