Implements reporting errors with message box for Windows GUI apps in apphost.
In apphost, if there's no console detected, error writing will be redirected to a buffer. Upon exit, if the error buffer is not empty, it will be showed on screen as a message box.
The error writer is propagated from apphost to hostfxr and thus to hostpolicy.

This solves the problem that GUI apps which don't have console die silently without reporting any errors if there are issus during host execution.