Change the algorithm to look for the GUI bit in PE header instead of trying to open console.
The previous version which used console stream as an indicator of a GUI app can lead to issues in cases where there's no console even for a "console" app - for example Windows Services are like this, they are compiled as console apps but run without attached console.

So to overcome this, this change modifies the algorithm to look for the GUI bit in the PE header directly. It will try to show a message box only for apps which are marked as GUI and thus should run in environment where showing GUI is OK.