Fix registry access on legacy Windows
The bug was that RRF_SUBKEY_WOW6432KEY is only supported on Win10. On legacy Windows it has no effect and 64bit registry is accessed.

The fix is to use RegOpenKeyEx which takes the access right parameter which can specify KEY_WOW64_32KEY. This makes 64bit app access the 32bit registry and this one is supported on anything above Windows 2000 (excluded).

Test is already in place.