Fix lookup of coreclr.dll in .deps.json to not rely on relative paths
The existing code expects the .deps.json record for coreclr.dll (and clrjit.dll) to have at least one directory separator, since that was always the case up until now.
With recent changes to targeting packs, these records are just file names, and the above logic breaks.

Fix the logic by lookup at the resolved absolute path, which is guaranteed to have a directory separator before the library name (we need the separator to match whole file names).

Added a test which validates the new scenario.