Add dotnet/versions publish as hosted job
If dotnet/versions publish is enabled in the main finalize/publish job, it fails because VersionTools is an old version. The reason for the old VersionTools is unknown. To narrow it down or work around it, use a hosted agent to ensure machine state isn't a factor, such as previous builds' tooling carrying over somehow.