Simplify computation of oldest framework reference
Instead of passing the list of oldest references all the way down to the code which performs roll forwards, simply assign to it when we enumerate reerences for the first time.
The only difference is that in case of a failure the oldest reference will be filled while before it would not be fully up to date.