Disable using clang 3.9 for linux ARM cross build
The clang 3.9 was found to be generating broken code when cross
building for arm/armel long time ago for coreclr repo builds.
It was generating e.g. incorrect code for interlocked operations.
Now with a recent change in core-setup (adding linking of libatomic.a),
it started to generate broken code in core-setup too (leading to
SIGILL in libhostpolicy.so).
This change disables usage of clang 3.9 for ARM cross building, like
we do in the coreclr repo.