Fix framework ordering if graphs
If both the app and framework (for example ASP.NET) have dependency on the root Microsoft.NETCore.App and these references are for different versions the product may not find `hostpolicy`.

The problem happens if the framework (ASP.NET for example) has a reference to lower (but still compatible) version to what the app has. In such case the app will resolve the framework and put it into the list. Then we process ASP.NET framework which gets added to the end of that list as well. And then we process references for ASP.NET, we find that the reference to NETCore.App is compatible with the one we already have (soft-roll-forward). In this case the NETCore.App should be moved to the end of the list (the list is ordered such that the root framework should be last), but the code used to only do that if the versions were an exact match.

The fix is to always move the dependent framework to the last position.

Tests we already added as part of the framework resolution test improvement, but commented out as they were failing. Just uncommenting the tests.