Implement more restrictive merge of roll forward
If LatestMinor and Major references are merged, previously the effective reference would be LatestMinor. Bu that goes against the Major semantics of "pick closest".
This change is modifying the outcome of that merge to Minor. This produces the more restrictive version selection (lower versions are prefered).

This behavior is still being discussed in dotnet/core-setup#5870 and may change.

Added tests for merging roll forward settings.