Implement two timeout approach to detect hangs.
As suggested by @danmosemsft the waiting now succeeds if the window is detected in the specified timeout. If it is detected but it took longer than that (up to 3 * timeout), the test fails but notes that it did get the window. If it takes longer than 3 * timeout, the test just fails.