Fix pal::get_own_executable_path() with multiple hardlinks on FreeBSD
The KERN_PROC_PATHNAME mib used to implement this function does not
work the way we expect it to work when there are multiple hardlinks to
an executable: it's not guaranteed to receive the actual path of the
hardlink that has been used to execute that process.

Since using this mib is the canonical way, continue doing so, but only
if the number of hardlinks is 1.

Otherwise, fallback to a crude hack using dladdr(), asking information
on the `main` symbol, guaranteed to be in the main executable.  On some
BSDs (such as OpenBSD), this doesn't return the real path of the
executable, but on FreeBSD 12 at least, it does, so we can rely on
this.  This method will fail if the main binary has been statically
built.

Fixes #4742