Improve apphost missing fxr error message with self registered location
With the recent changes apphost now looks for shared framework in 3 potential places: the default global location, the globally registered location and the DOTNET_ROOT path.
The error message didn't mention the globally registered location, so add that.

Required a bit of refactoring in the pals to avoid code duplication.