Add COLD_FUNCTION, PURE_FUNCTION, and NO_INLINE function attributes to PAL
These attributes are supported by GCC and Clang and allow functions to
be, respectively, marked as "cold" (e.g.  code is kept away from any
hot path, it's not marked as a potentially interesting optimization
target, modifies behavior regarding branch prediction, etc.), "pure"
(allowing some optimizations for repeated calls) or "noinline" (so that
the compiler will never consider those functions to be inlined).

The idea is to use them in the tracing subsystem:
- Errors are rare, tracing them should not use (valuable) I$.
- Checking the current verbosity level multiple times from the same
  function do not need to incur in a function call every time.