Lazily evaluate tracer arguments
Introduce TRACE_ERROR(), TRACE_WARNING(), TRACE_INFO(), and
TRACE_VERBOSE() macros, that supersede similarly-named functions in the
trace namespace.

Using these macros will ensure that the arguments are only evaluated in
the expected verbosity levels.  Some arguments create temporary objects
that are only visible in high verbosity levels, creating useless churn,
that often show up in profilers.

Another benefit of this is that error conditions are very rare, so they
shouldn't be using instruction cache real-state, interfering with
prefetchers, branch predictors, and whatnot.  So all error traces are
delegated to a self-calling lambda, that is marked as "cold" (so far away
from the happy path) and not inlined.

Further improvements to the tracing subsystem may use something other
than stdio/printf to format, possibly using variadic templates.  This
is feasible without a lot of changes because throughout corehost, only
the "%s", "%x", and "%d" conversion specifiers are used, so it's easier
to implement something tailored to the current use.  Without using C++17,
however, this is overly complicated and this solution using macros is
cleaner.