WIP: Add AppHost Binary transformatuin Utilities to HostModel
The code that transforms AppHost binary in order to update the
path to the App.dll binary lives in multiple repositories
(SDK, CLI and partly in core-setup tests).

This change adds the code to HostModel library, so that other
copies can be removed.

This also facilitates implementing a similar mechanism to
locate the offset of the bundle header within the binary
(so that the bundle headers are not required to live at the end
of the file).