Fix Alpine rootfs build
Alpine has moved to a new version of LLDB (8) and its dependency LLVM
libraries are now in the edge/main branch. This adds edge/main branch to
the apk add command commandline so that the dependencies can be found.

This is a copy of the same change made in dotnet/arcade. We should eventually
move core-setup to get the eng folder from arcade, but for now, we need
a quick solution.