Skip pal::realpath() error logging while probing directories
Both paths calling pal::realpath() and check its return value to see if
the probed paths actually exist in the filesystem; they do have logic
to handle those paths not existing, and printing error messages while
doing so isn't helpful for the user.

This happens when an application is published with one user but is
executed with another user -- for instance, when using containers (see
issue #4038 for details).

Fixes #4038.