Fixes couple of places in the bundle runner where we might leak memory.
Note that in reality this is a non-issue: The leaks are only present if something fails and in case of failure the process will immediately exit anyway.

I did this mostly for my own education on how to write this correctly. And since I had the changes read, we might as well merge it.