Add root "shared" dir to WindowsDesktop zip
Also fix a small bug where both the new and old infrastructure were creating a zip for the NETCoreApp sfx. This wasn't causing a problem in parallel because they are ordered, but it made it unclear to a reader which infra was responsible for the zip.

(cherry picked from commit 732ef42a49e82b1bfbe1363fa6799b34d7e8ef34)