[release/3.0] HostWriter: Retry updates on IO error (#7617) (#7638)
* HostModel: Retry updates on IO error (#7617)

The HostModel implements services for updating the AppHost DLL
- Write App-dll name
- Set GUI/CUI bit
- Read/Write BundleHeader marker

These updates involve multiple file open/close operations. An Antivirus scanner may intercept in-between and cause these operations to fail with IO-Error.

So, this change adds a few retries to the IO operations, similar to
https://github.com/dotnet/core-setup/pull/7135

On Win32 PE files, the SDK copies resources on the AppHost binary from the intermediate assembly.

This update is performed using native Win32 operations. This change also adds retries to the resource update, if the operation fails because the file is locked (say because of AntiVirus scan).

The Retry logic is factored out to the RetryUtil class. Currently the HostWriter is the only client for the class. But it is made public because other components (like Bundler) may use it in future.