Unset $DOTNET_ROOT environment variable for tests that don't require it
If this variable is set in your environment, running `build.sh -t` will
pass it to the test harness, causing these tests to fail.

Teach Command.EnvironmentVariable() to handle `null` as an empty string,
so that if the DotNetRoot() extension method is called with a `null`
parameter, the $DOTNET_ROOT and $DOTNET_ROOT(x86) environment variables
are essentially unset as far as pal::getenv() is concerned.