Double publish artifacts in cases of stable builds (#8443)
Stable assets do not publish to stable directories to avoid overwrite issues. But, the dotnet-install script expects the runtime version and directory to match, which they wouldn't. To deal with this, in cases of stable builds, copy all the stable branded bits to files with non-stable names.  Thus, the suffixed directory will contain the stable file names (e.g. 3.0.1) as well as suffixed file names that work prior to release day. On release day, the directory is copied or renamed and the dotnet-install script will work with the official versions.