Set InitialAssetsLocation as well as location (#8969)
This repo was only setting Location. This setting is being deprecated for InitialAssetsLocation. Without setting InitialAssetsLocation, the Location parameter is used to populate the initial locations of each asset. This means that Maestro was thinking that all assets in core-setup, coming out of the build, were sitting in the dotnet-core feed, which is not true.