[release/3.1] Single-File: Unix: Extract to users-specific temp directory
** Issue

https://github.com/dotnet/runtime/issues/3846

** Customer Scenario

On some Unix systems, multiple users cannot use single-file apps on the same machine.

** Problem

In .net core 3, single file apps run by extracting the bundled contents to a temp directory.
The extraction directory is machine specific, and can be set through DOTNET_BUNDLE_EXTRACT_BASE_DIR environment variable.

When this setting is not configured, the host tries to use certain default directories.
On windows, extraction is within %TMPDIR%, which is user specific.
On Unix systems $TMPDIR/.net if set, which may be user specific (ex: MAC)
Otherwise, the extraction directory is within /var/tmp/ or /tmp/ which is common to all users, and may be locked by a specific user on first creation.

** Solution

This change fixes this issue by defaulting the extraction base directory in Unix systems to
<temp-dir>/.net/<user-ID> , where
<temp-dir>/.net/ has permission 0777, and
<temp-dir>/.net/<user-ID>/ has permission 0700.

** Risk

Low, scenario is contained, change is small.

** Master Branch

https://github.com/dotnet/runtime/pull/2329